Update getwords.py
Move a significant amount of logic to the python script. The new logic will reject a high number of comments, but this is to increase the likeliness that the comment is in English, not spam, and contains full thoughts. This change also will speed up the overall process and cause less system load.